+++ src/pokemon_stats.c
@@ -1,7 +1,7 @@
 #include "global.h"
 #include "pokemon_stats.h"
 
-u8 GetType1(u16 species, u8 variant) {
+u8 GetType1(u16 species, u16 variant) {
     if ((gVariantStats[species][variant].variantFlags & VARIANT_FLAG_TYPE) > 0) {
         return gVariantStats[species][variant].type1;
     }
@@ -10,7 +10,7 @@ u8 GetType1(u16 species, u8 variant) {
     }
 }
 
-u8 GetType2(u16 species, u8 variant) {
+u8 GetType2(u16 species, u16 variant) {
     if ((gVariantStats[species][variant].variantFlags & VARIANT_FLAG_TYPE) > 0) {
         return gVariantStats[species][variant].type2;
     }
@@ -19,11 +19,11 @@ u8 GetType2(u16 species, u8 variant) {
     }
 }
 
-u8 GetLevelUpLearnset(u16 species, u8 variant, u16 *learnset) {
+u8 GetLevelUpLearnset(u16 species, u16 variant, u16 *learnset) {
     u8 count = 0;
     int i;
     u8 baseIndex = 0;
-    u8 variantIndex = 0;
+    u16 variantIndex = 0;
     u16 baseLevel;
     u16 variantLevel;
